third_party/WebKit/LayoutTests/shadow-dom/imperative-api.html

<!DOCTYPE html>
<!--
Tests for Imperative Shadow DOM Distribution API.
See https://crbug.com/869308
-->
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="host">
  <div id="child1"></div>
  <div id="child2"></div>
  <div id="child3"></div>
</div>
<div id="host1">
</div>
</div>
<div id="host2">
</div>
</div>
<div id="host3">
</div>
<div id="child4"></div>
<div id="host4">
  <div id="child5"></div>
</div>
</div>
<div id="host5">
</div>

<script>
const host = document.querySelector('#host');
const host1 = document.querySelector('#host1');
const host2 = document.querySelector('#host2');
const host3 = document.querySelector('#host3');
const child1 = document.querySelector('#child1');
const child2 = document.querySelector('#child2');
const child3 = document.querySelector('#child3');
const child4 = document.querySelector('#child4');
const child5 = document.querySelector('#child5');
const shadow_root = host.attachShadow({ mode: 'open', slotting: 'manual' });
const shadow_root1 = host4.attachShadow({ mode: 'open' });
const shadow_root2 = host5.attachShadow({ mode: 'open', slotting: 'manual' });
const slot1 = document.createElement('slot');
const slot2 = document.createElement('slot');
const slot3 = document.createElement('slot');
const slot4 = document.createElement('slot');
const slot5 = document.createElement('slot');
const slot6 = document.createElement('slot');
shadow_root.appendChild(slot1);
shadow_root.appendChild(slot2);
shadow_root.appendChild(slot3);
shadow_root.appendChild(slot6);
shadow_root1.appendChild(slot4);

test(() => {
  assert_not_equals(host1.attachShadow({ mode: 'open', slotting: 'manual' }),
                    null, 'slotting manual should work');
  assert_not_equals(host2.attachShadow({ mode: 'open', slotting: 'auto' }),
                    null, 'slotting auto should work');
  assert_throws(new TypeError(), () => {
    host3.attachShadow({ mode: 'open', slotting: 'exceptional' })},
                'others should throw exception');
}, 'attachShadow can take slotting parameter');

test(() => {
  assert_array_equals(slot2.assignedElements(), []);
  assert_equals(child1.assignedSlot, null);

  slot2.assign([child1]);

  assert_array_equals(slot2.assignedNodes(), [child1]);
  assert_equals(child1.assignedSlot, slot2);

  slot1.assign([child2,child1]);
  assert_array_equals(slot1.assignedNodes(), [child1,child2]);
  assert_array_equals(slot2.assignedNodes(), []);
  assert_equals(child1.assignedSlot, slot1);
  assert_equals(child2.assignedSlot, slot1);

  slot1.assign([child2]);
  assert_array_equals(slot1.assignedNodes(), [child2]);
  assert_array_equals(slot2.assignedNodes(), [child1]);
}, 'assignedNodes/Slot can be used in manual slotting');

test(() => {
  slot3.assign([child4]);
  assert_array_equals(slot3.assignedNodes(), []);

  host.appendChild(child4);
  assert_array_equals(slot3.assignedNodes(), [child4]);
}, 'Nodes should be assigned to slot only when nodes are host\'s children');

test(() => {
  slot5.assign([child1]);
  assert_array_equals(slot5.assignedNodes(), []);

  shadow_root.insertBefore(slot5,slot1);
  assert_array_equals(slot5.assignedNodes(), [child1]);
}, 'Nodes should be assigned to a slot only when the slot is in host\'s shadowtree');

test(() => {
  slot1.assign([child2]);
  slot2.assign([child2]);
  assert_array_equals(slot2.assignedNodes(), []);

  shadow_root.insertBefore(slot2,slot1);
  assert_array_equals(slot2.assignedNodes(), [child2]);
}, 'A node should be assigned to a slot when the slot is inserted before the assigned slot');

test(() => {
  shadow_root.appendChild(slot1);
  shadow_root.appendChild(slot2);
  shadow_root.appendChild(slot6);
  slot1.assign([child2]);
  slot2.assign([child2]);
  slot6.assign([child2]);
  assert_array_equals(slot1.assignedNodes(), [child2]);

  shadow_root.insertBefore(slot6,slot2);
  slot1.remove();
  assert_array_equals(slot6.assignedNodes(), [child2]);
}, 'A slot should be assigned to a node even after the slot is inserted after the assigned slot, and the assigned slot are removed')

  test(() => {
  shadow_root.insertBefore(slot1,slot2);
  slot2.assign([child3,child3]);
  assert_array_equals(slot2.assignedNodes(), [child3]);

  slot1.assign([child3]);
  assert_array_equals(slot2.assignedNodes(), []);

  shadow_root2.appendChild(slot2);
  assert_array_equals(slot2.assignedNodes(), []);
}, 'Same nodes should be ignored');

test(() => {
  assert_array_equals(slot4.assignedElements(), [child5]);
  assert_equals(child5.assignedSlot, slot4);

  slot4.assign([]);
  assert_array_equals(slot4.assignedElements(), [child5]);
  assert_equals(child5.assignedSlot, slot4);
}, 'slotting API should not have any effect in auto mode');
</script>
